change: use github-flask instead of flask-dance
