+++ DEPENDENCIES
@@ -21,9 +21,9 @@ pypi/pypi/-/cryptography/42.0.5
 pypi/pypi/-/dnspython/2.6.1
 pypi/pypi/-/exceptiongroup/1.2.0
 pypi/pypi/-/flask/3.0.2
-pypi/pypi/-/flask-dance/7.1.0
 pypi/pypi/-/frozenlist/1.4.1
 pypi/pypi/-/ghp-import/2.1.0
+pypi/pypi/-/github-flask/3.2.0
 pypi/pypi/-/gojsonnet/0.20.0
 pypi/pypi/-/graphql-core/3.2.3
 pypi/pypi/-/greenlet/3.0.3
@@ -51,7 +51,6 @@ pypi/pypi/-/mkdocs-material/9.5.17
 pypi/pypi/-/mkdocs-material-extensions/1.3.1
 pypi/pypi/-/motor/3.4.0
 pypi/pypi/-/multidict/6.0.5
-pypi/pypi/-/oauthlib/3.2.2
 pypi/pypi/-/odmantic/1.0.1
 pypi/pypi/-/packaging/24.0
 pypi/pypi/-/paginate/0.5.6
@@ -80,7 +79,6 @@ pypi/pypi/-/redis/4.6.0
 pypi/pypi/-/referencing/0.34.0
 pypi/pypi/-/regex/2023.12.25
 pypi/pypi/-/requests/2.31.0
-pypi/pypi/-/requests-oauthlib/2.0.0
 pypi/pypi/-/rpds-py/0.18.0
 pypi/pypi/-/six/1.16.0
 pypi/pypi/-/sniffio/1.3.1
@@ -90,7 +88,6 @@ pypi/pypi/-/tomli/2.0.1
 pypi/pypi/-/typing-extensions/4.10.0
 pypi/pypi/-/url-normalize/1.4.3
 pypi/pypi/-/urllib3/2.2.1
-pypi/pypi/-/urlobject/2.4.3
 pypi/pypi/-/watchdog/4.0.0
 pypi/pypi/-/werkzeug/3.0.2
 pypi/pypi/-/wsproto/1.2.0
