+++ otterdog/webapp/templates/auth/login.html
@@ -28,7 +28,7 @@
 
         <div class="text-center">
           <div class="text-center mx-auto">
-            <a href="{{url_for('github.login')}}" class="btn btn-github btn-lg">
+            <a href="{{url_for('auth_blueprint.login_github')}}" class="btn btn-github btn-lg">
                 <i class="fab fa-github"></i> GitHub
             </a>
           </div>
