+++ otterdog/webapp/templates/includes/navigation.html
@@ -26,12 +26,12 @@
                 </div>
             </a>
             <div class="dropdown-menu dropdown-menu-sm dropdown-menu-right">
-              <a href="{{ url_for('github.logout') }}" class="dropdown-item">
+              <a href="{{ url_for('auth_blueprint.logout') }}" class="dropdown-item">
                 {{ current_user.username }} - Logout
               </a>
             </div>
             {% else %}
-            <a href="{{ url_for('github.login_view') }}" class="dropdown-item">
+            <a href="{{ url_for('auth_blueprint.login_view') }}" class="dropdown-item">
               <i class="fas fa-user mr-2"></i> Login
             </a>
             {% endif %}
