+++ pyproject.toml
@@ -64,7 +64,7 @@ pydantic             = "^2.6"
 odmantic             = "^1.0.1"
 quart-auth           = "^0.9"
 quart-redis          = "^2.0"
-flask-dance          = "^7.1"
+github-flask         = "^3.2"
 aiohttp-client-cache = {version="^0.11.0", extras=["redis"]}
 ariadne              = "^0.22"
